structure fmgc components
