+++ src/fmgc/src/components/FmgcComponent.ts
@@ -1,7 +1,4 @@
 export interface FmgcComponent {
-
     init(): void;
-
     update(deltaTime: number): void;
-
 }
