+++ src/fmgc/src/components/fms-messages/index.ts
@@ -1,0 +1,1 @@
+export { FmsMessages } from './FmsMessages';
